Remove the native validation from the forms
React Hook Form is used with a resolver (Yup) to validate the form. Unfortunately, RHF is not able to validate individual fields when using one. This still worked well when the user would manually change the value of an input, but with the `<TagGroup />` component (#80), we need to trigger changes to the field from the outside and trigger validation on that field only, otherwise errors could be triggered on fields the user haven't had a look yet.