Fix bug preventing sign out
Previously, to sign out, the user would be redirected to a page where a `useEffect` would delete the session. Unfortunately, that same page would also request the account information via the header of the static layout. This resulted in a race condition and the API would not delete the session consistently.